Add helpers for sheet index updates and improve warnings
Introduced ComputeEndIndex and GenerateUpdateSheetIndex helpers in GoogleRequestHelpers with corresponding unit tests. Removed redundant integration tests. Updated GoogleSheetManager to add a warning message if moving the default sheet fails, instead of silently ignoring the error.
